requirements.txt
streamlit
pandas
matplotlib
plotly
pillow
openpyxl
openai